Split selfplay utilities into smaller modules
